separate the asset server from the rest of the servers
